refactor(agents): delete the graph branch, keep `applyRunPolicy` (T-A3.4)
`executeAgentGraph`, `resolveAgentGraph` and `runWorkflowAsAgent` had one
importer each — `agent.ts` — plus a re-export in `index.ts`. With `Agent`
retired they have none, so `execute-agent-graph.ts` and `workflow-agent.ts` go.

`applyRunPolicy` survives on the same evidence: `evals/graph-e2e-eval.ts`
imports it, and it is a graph transform rather than execution wiring — it
stamps provider, model and prompt onto a planned graph. It moves to
`run-policy.ts` so what remains is a file about the thing that is left, not the
husk of the one that went, and gets its own test now that its old coverage went
with the deleted module.

Two catalogues that named the deleted files are corrected in the same change:
`packages/execution`'s README surface table and its preflight-audit
allowlist, which carried an entry per file for how each propagated a refusal.
Left alone, the audit would have kept vouching for modules that no longer exist.

Part of A3 in docs/plans/agent-system-improvements.md (T-A3.4).
