+++ packages/execution/README.md
@@ -37,7 +37,7 @@ execution surface — not part of this facade's contract).
 | `packages/websocket/src/http-api.ts:771` (async HTTP run) | — | — | params only | — | — | — | — | **Migrated.** The route calls `runWorkflow` from `@nodetool-ai/execution/service`, so hydration, executor resolution, cancellation, timeout and `Job` persistence are all the facade's now. The one thing it still owns is the environment: `environment: () => getWorkflowRuntimeEnvironment(options)`, passed lazily so a nonexistent workflow 404s before the runtime bootstraps. That bootstrap stays **out of scope** (it also builds the registry and connects the Python bridge — infra construction, not execution semantics) |
 | `packages/websocket/src/mcp-server.ts` (MCP tool: run workflow) | — | — | — | — | — | — | — | **Gone.** The mount no longer runs workflows; `registerAgentMcpTools` exposes the agents package's capability tools instead, and the file imports nothing from `@nodetool-ai/kernel` |
 | `packages/core-nodes/src/nodes/run-inner-graph.ts:131` (`RunInnerGraph` / `WorkflowNode`) | `Graph.loadFromDict` via `context.resolveNodeType` when set, else `withExplicitNodeFlags` | injected `resolveExecutor` from the *parent* run's `context.resolveExecutor` | params from node properties | none (bounded by the parent run's lifetime) | none | none (child of an already-persisted parent run) | maps runner output keys back to the sub-graph's declared output names (`outputKeyMap`) | **out of scope**: a node's own implementation spawning a *child* runner inside an already-running parent node — not a top-level surface entry point. Revisit only if a future task generalizes nested execution |
-| `packages/agents/src/execute-agent-graph.ts:122` (`executeAgentGraph`) | the facade's, via `ExecutionSession.create({resolveNodeType})`; `applyRunPolicy` first stamps provider/model/prompt onto the planned `GraphData` | the facade's, from the passed `registry` plus `createGraphNodeTypeResolver(registry).resolveNodeType` | `params: opts.inputs ?? {}` | `session.cancel("cancelled")`, driven by `opts.signal` and by an early-`break` consumer | none | none | none | **Migrated.** `AgentWorkflowRunner` was flattened into `executeAgentGraph` (#4907), which drives `ExecutionSession` instead of constructing a `WorkflowRunner`. `applyRunPolicy` stays agent-owned — it is a graph transform, not execution wiring. Persistence stays absent: this streams into a chat turn, not a job |
+| `packages/agents/src/execute-agent-graph.ts:122` (`executeAgentGraph`) | — | — | — | — | — | — | — | **Gone.** The retired `Agent` graph branch was its only caller, so it was deleted with it; `applyRunPolicy`, a graph transform rather than execution wiring, survives in `packages/agents/src/run-policy.ts` for the `graph-e2e` eval |
 | `packages/dsl/src/core.ts:399` (`run()`, the TS DSL entry point) | none — DSL nodes carry flags from generated metadata; wrapped in `withExplicitNodeFlags` directly | 4-branch: `opts.registry` → `NodeRegistry.global` → `builtinRegistry` → Python bridge → throw (a superset of the 3-branch pattern — DSL graphs can mix three TS registries) | params only | none | none | none | none | **out of scope for A2, revisit for A3+**: the DSL's multi-registry resolver is a genuinely different shape (3 TS registries, not 1) that this package's single-`registry` option doesn't yet cover. Noted as a gap, not silently generalized |
 | `packages/workflow-runner/src/run.ts:118` (`runWorkflow`, browser + `nodetool run` shared core) | `hydrateGraphNodeFlags` | `opts.registry.resolve(node)` only (no Python-bridge branch — this path targets the browser, which has no Python bridge) | params only, **but** already an `AsyncGenerator<ProcessingMessage, RunResult>` with `AbortSignal`-based `cancelRun()` — this file is the closest existing prior art to `ExecutionSession`'s streaming/cancel shape | `AbortSignal` → `runner.cancel()`, mirrors this package's `cancel()` | none built in (caller's `AbortSignal` can carry one) | none (caller's concern) | none | **not migrated in A2** — its generator-based public API predates this package and has its own consumers (browser workflow runner, `nodetool run`). Cataloged here as the pattern this package's `messages`/`cancel()` API generalizes; a follow-up task can make it a thin wrapper over `ExecutionSession` |
 | `packages/workflow-runner/e2e/src/browser-entry.ts:151` | same pattern as `run.ts` (it's the e2e harness's copy of the browser runner) | `opts.registry.resolve` | params only | `AbortSignal` | none | none | none | **out of scope**: test harness, not a product surface |
